requirement.txt
openai
pandas
pillow 
python-dotenv
stability-sdk